Refactored User Authentication Refactored AdministrativeService Added ReadingService
Functionality Completed are

authentication
authorization
add book
remove book
searchBookByName

Functionality In Progress are
borrowBook

+++ src/main/service/AdministrativeService.java
@@ -1,70 +1,42 @@
 package main.service;
 
-import main.model.Book;
-import main.repository.AuthorRepo;
+import main.model.*;
 import main.repository.BookRepo;
-import main.repository.PublisherRepo;
 
 import java.sql.SQLException;
+import java.util.ArrayList;
 import java.util.List;
 
 public class AdministrativeService {
 
     private final BookRepo bookRepo;
-    private final PublisherRepo publisherRepo;
-    private final AuthorRepo authorRepo;
 
-    public AdministrativeService(BookRepo bookRepo, AuthorRepo authorRepo, PublisherRepo publisherRepo) {
+    public AdministrativeService(BookRepo bookRepo) {
         this.bookRepo = bookRepo;
-        this.authorRepo = authorRepo;
-        this.publisherRepo = publisherRepo;
     }
 
-    public int addBook(String name, List<String> authorNames, String publisherName, int noOfCopies) throws SQLException {
-
-        Integer[] authorIds = populateAuthorIds(authorNames);
-        int publisherId = populatePublisherId(publisherName);
-
-        Book book = new Book(name, authorIds, publisherId, noOfCopies);
-        int addedBookId = -1;
-
-        try {
-            addedBookId = bookRepo.addBook(book);
-        } catch (SQLException e) {
-            e.printStackTrace();
-        } finally {
-            return addedBookId;
-        }
-
-    }
-
-    private int populatePublisherId(String publisherName) throws SQLException {
-        int publisherId = publisherRepo.retrieveId(publisherName);
-        if (publisherId == -1)
-            publisherId = publisherRepo.create(publisherName);
-        return publisherId;
-    }
-
-    private Integer[] populateAuthorIds(List<String> authorNames) throws SQLException {
-        Integer[] authorIds = new Integer[authorNames.size()];
-
-        int i = 0;
+    public Book addBook(User user, String name, List<String> authorNames, String publisherName, int noOfCopies) throws Exception {
+        if (!user.isAuthorized(Permission.ADD_BOOK)) throw new Exception("User Not Authorized");
+        List<Author> authors = new ArrayList<>();
         for (String authorName : authorNames) {
-            int id = authorRepo.retrieveId(authorName);
-            if (id != -1)
-                authorIds[i] = id;
-            else
-                authorIds[i] = authorRepo.create(authorName);
-            i++;
+            authors.add(new Author(authorName));
         }
-        return authorIds;
+        Publisher publisher = new Publisher(publisherName);
+
+        Book book = new Book(name, authors, publisher, noOfCopies);
+        Book addedBook = bookRepo.addBook(book);
+        return addedBook;
+
     }
 
-    public boolean removeBook(Book book) throws SQLException {
+    public boolean removeBook(User user, Book book) throws Exception {
+        if (!user.isAuthorized(Permission.REMOVE_BOOK)) throw new Exception("User not authorized for this operation");
         return bookRepo.delete(book);
     }
 
     public List<Book> searchBookByName(String name) throws SQLException {
-        return bookRepo.retrieveByName(name);
+        return bookRepo.searchBookByName(name);
     }
+
+
 }
